# openapi.json
{
  "components": {
    "schemas": {
...
      "ApiHistoryResponse": {
        "properties": {
          "cache_rows": {
            "title": "Cache Rows",
            "type": "integer"
          },
          "clicks": {
            "title": "Clicks",
            "type": "integer"
          },
          "items": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ClickItem"
                },
                {
                  "$ref": "#/components/schemas/CacheItem"
                }
              ]
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "since": {
            "title": "Since",
            "type": "string"
          }
        },
        "required": [
          "items",
          "clicks",
          "cache_rows",
          "limit",
          "since"
        ],
        "title": "ApiHistoryResponse",
        "type": "object"
      },
      "ApiStatsResponse": {
        "additionalProperties": true,
        "description": "Dashboard aggregates; stats.py owns the exact shape, extras allowed.",
        "properties": {
          "days": {
            "title": "Days",
            "type": "integer"
          }
        },
        "required": [
          "days"
        ],
        "title": "ApiStatsResponse",
        "type": "object"
      },
      "CacheInvalidateResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "integer"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "CacheInvalidateResponse",
        "type": "object"
      },
      "CacheItem": {
        "additionalProperties": true,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "integer"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "integer"
          },
          "hits": {
            "title": "Hits",
            "type": "integer"
          },
          "kind": {
            "const": "cache",
            "title": "Kind",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "query_hash": {
            "title": "Query Hash",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "kind",
          "created_at",
          "query_hash",
          "query",
          "expires_at",
          "hits",
          "size_bytes"
        ],
        "title": "CacheItem",
        "type": "object"
      },
...
      "ClickItem": {
        "additionalProperties": true,
        "properties": {
          "clicked_at": {
            "title": "Clicked At",
            "type": "integer"
          },
          "kind": {
            "const": "click",
            "title": "Kind",
            "type": "string"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "query_hash": {
            "title": "Query Hash",
            "type": "string"
          },
          "result_id": {
            "title": "Result Id",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "clicked_at",
          "query_hash",
          "query",
          "result_id",
          "url",
          "title",
          "source"
        ],
        "title": "ClickItem",
        "type": "object"
      },
      "ClickRequest": {
        "properties": {
          "query_hash": {
            "title": "Query Hash",
            "type": "string"
          },
          "result_id": {
            "title": "Result Id",
            "type": "string"
          },
          "source": {
            "default": "web-ui",
            "enum": [
              "web-ui",
              "mcp"
            ],
            "title": "Source",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "query_hash",
          "result_id",
          "url"
        ],
        "title": "ClickRequest",
        "type": "object"
      },
      "ClickResponse": {
        "properties": {
          "click_id": {
            "title": "Click Id",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "click_id"
        ],
        "title": "ClickResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "oxe",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/history": {
      "get": {
        "description": "JSON API for the SPA /history page: merged user activity view.\n\nParams: since=24h|7d|30d|all (default all), q=substring filter,\nlimit=1..200 (default 50), kind=all|clicks|cache (default all).\nReturns {items, clicks, cache_rows, limit, since} sorted newest\nfirst; each item has a `kind` field (\"click\" | \"cache\").",
        "operationId": "api_history_api_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "24h",
                "7d",
                "30d",
                "all"
              ],
              "title": "Since",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "all",
              "enum": [
                "all",
                "clicks",
                "cache"
              ],
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api History"
      }
    },
    "/api/stats": {
      "get": {
        "description": "JSON API for the SPA /dashboard page: aggregates from search_log\n(per-day, hit rate, latency percentiles, top/zero-result queries,\nclient split) plus cache stats. Params: days=1..90 (default 14).",
        "operationId": "api_stats_api_stats_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 14,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiStatsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Api Stats"
      }
    },
    "/assets/{name}": {
      "get": {
        "description": "Serve built SPA assets from the UI dist dir (404 when absent).",
        "operationId": "ui_assets_assets__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Assets"
      }
    },
    "/cache": {
      "get": {
        "operationId": "ui_cache_cache_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "include_expired",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Expired",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Cache"
      }
    },
    "/cache/invalidate": {
      "post": {
        "operationId": "cache_invalidate_cache_invalidate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CacheInvalidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cache Invalidate"
      }
    },
...
    "/click": {
      "post": {
        "operationId": "ui_click_click_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClickRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClickResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Click"
      }
    },
...
    "/history": {
      "get": {
        "operationId": "ui_history_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui History"
      }
    },
...
    "/row/{key}": {
      "get": {
        "operationId": "ui_row_row__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Row"
      }
    },
    "/row/{key}/delete": {
      "post": {
        "operationId": "ui_row_delete_row__key__delete_post",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ui Row Delete"
      }
    },
...
  }
}